+++ src/Sentry.Unity/UnityEventProcessor.cs
@@ -41,7 +41,7 @@ internal class UnityEventProcessor :
             // that it got added last or that there was not an integration added at a later point
             PopulateSdkIntegrations(sentryEvent.Sdk);
             // TODO revisit which tags we should be adding by default
-            sentryEvent.SetTag("unity.is_main_thread", MainThreadData.IsMainThread().ToTagValue());
+            sentryEvent.SetTag("unity.is_main_thread", _sentryOptions.IsMainThread().ToTagValue());
         }
         catch (Exception exception)
         {
@@ -51,7 +51,7 @@ internal class UnityEventProcessor :
 
     private void PopulateDevice(Device device)
     {
-        if (!MainThreadData.IsMainThread())
+        if (!_sentryOptions.IsMainThread())
         {
             return;
         }